Set all RESET_* environment variables to 0 by default for safety
- And other tidy ups
- ref_type in pushdata table should be a tinyint

+++ conduit_lib/database/scylladb/tables.py
@@ -112,7 +112,7 @@ class ScyllaDBTables:
                     pushdata_hash blob,
                     tx_hash blob,
                     idx bigint,
-                    ref_type smallint,
+                    ref_type tinyint,
                     PRIMARY KEY (pushdata_hash, tx_hash, idx, ref_type)
                 );
                 """
